[MISC] Speed up vllm import
